# specs/001-b2b2c/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 商家入驻管理 API
  description: B2B2C 平台商家入驻系统 RESTful API 规范
  version: 1.0.0
  contact:
    name: API Support
    email: api@grainadmin.com
servers:
  - url: https://api.grainadmin.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Development
security:
  - BearerAuth: []
...
paths:
  /applications:
    post:
      tags: [applications]
      summary: 创建商家入驻申请
      description: 商家创建新的入驻申请（初始状态为 DRAFT）
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: 申请创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - BearerAuth: []

    get:
      tags: [applications]
      summary: 查询商家申请列表
      description: |
        查询商家申请列表（支持分页和筛选）
        - 商家用户：只能查询自己的申请
        - 平台管理员：可查询所有申请
      operationId: listApplications
      parameters:
        - name: status
          in: query
          description: 申请状态筛选
          schema:
            type: string
            enum: [draft, pending, under_review, require_supplement, approved, rejected]
        - name: page
          in: query
          description: 页码（从 1 开始）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /audits:
    post:
      tags: [audits]
      summary: 创建审核记录
      description: |
        审核员对商家申请进行审核操作（通过/拒绝/要求补充材料）
      operationId: createAudit
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditRequest'
      responses:
        '201':
          description: 审核记录创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /merchants/{merchant_id}:
    get:
      tags: [merchants]
      summary: 查询商家账户信息
      description: |
        查询商家账户详细信息
        - 商家用户：只能查询自己的账户
        - 平台管理员：可查询所有商家
      operationId: getMerchant
      parameters:
        - name: merchant_id
          in: path
          required: true
          description: 商家编号（如 M20250930001）
          schema:
            type: string
            pattern: '^M\d{11}$'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [merchants]
      summary: 更新商家信息
      description: |
        商家更新自己的基本信息（联系方式、经营类目等）
      operationId: updateMerchant
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    CreateApplicationRequest:
      type: object
      required:
        - business_name
        - unified_social_credit
        - legal_person_name
        - legal_person_id_number
        - contact_name
        - contact_phone
        - contact_email
        - business_categories
      properties:
        business_name:
          type: string
          maxLength: 200
          example: "测试科技有限公司"
        unified_social_credit:
          type: string
          pattern: '^[0-9A-HJ-NPQRTUWXY]{2}\d{6}[0-9A-HJ-NPQRTUWXY]{10}$'
          example: "91110108MA01234567"
        legal_person_name:
          type: string
          maxLength: 100
          example: "张三"
        legal_person_id_number:
          type: string
          pattern: '^\d{17}[\dXx]$'
          example: "110101199001011234"
        contact_name:
          type: string
          maxLength: 100
          example: "李四"
        contact_phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
          example: "13800138000"
        contact_email:
          type: string
          format: email
          example: "contact@example.com"
        business_categories:
          type: array
          minItems: 1
          items:
            type: string
          example: ["食品", "日用品"]
...
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_name:
          type: string
        unified_social_credit:
          type: string
        legal_person_name:
          type: string
        contact_name:
          type: string
        contact_phone:
          type: string
        contact_email:
          type: string
        business_categories:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, pending, under_review, require_supplement, approved, rejected]
        submitted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ApplicationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
        total:
          type: integer
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        page_size:
          type: integer
          description: 每页数量
    CreateAuditRequest:
      type: object
      required:
        - application_id
        - action
        - result
      properties:
        application_id:
          type: string
          format: uuid
        action:
          type: string
          enum: [start_review, approve, reject, request_supplement]
        result:
          type: string
          enum: [approved, rejected, pending_supplement, in_progress]
        comment:
          type: string
          description: 审核意见（reject 和 request_supplement 时必填）
    AuditResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        auditor_id:
          type: string
          format: uuid
        auditor_name:
          type: string
        action:
          type: string
        result:
          type: string
        comment:
          type: string
        created_at:
          type: string
          format: date-time
...
    MerchantResponse:
      type: object
      properties:
        merchant_id:
          type: string
          example: "M20250930001"
        username:
          type: string
        level:
          type: string
          enum: [normal, vip, diamond]
        status:
          type: string
          enum: [active, frozen, closed]
        application_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    UpdateMerchantRequest:
      type: object
      properties:
        contact_phone:
          type: string
        contact_email:
          type: string
          format: email
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误详细信息
        details:
          type: object
          description: 额外错误详情
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'